Update BulkIngest example to work with 2.0
Update the bulkIngest example to work correctly with Accumulo 2.0.x.
Minor updates made to the documentation. The classes were refactored so
the example will work correctly with 2.0 changes. Generation of test
data was moved from SetupTable to BulkIngestExample (as in 1.10.x
version). Prior to change, necessary data was not correctly being written
to HDFS thereby preventing the BulkIngestExample class from finding
information needed to generate the data.